requirements.txt
flask
mysql
redis
